requirements.txt
numpy
tensorflow==1.7.0
pillow
scipy
opencv-python
requests
flask
scikit-learn==0.19.2
